Create AppLayout wrapper
To wrap every page in a layout component.
