+++ gatsby-ssr.tsx
@@ -1,0 +1,1 @@
+export { wrapPageElement } from "./gatsby-browser"
