--- src/@chakra-ui/gatsby-plugin/theme.ts
@@ -1,4 +1,0 @@
-import { extendTheme } from "@chakra-ui/react"
-import ThresholdTheme from "../../theme"
-
-export default extendTheme(ThresholdTheme)
