+++ src/components/Navbar/DesktopNav/NavLink.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from "react"
+import { FC } from "react"
 import { Link } from "@chakra-ui/react"
 import { Label1 } from "../../Typography"
