+++ src/components/index.ts
@@ -1,1 +1,6 @@
 export * from "./Footer"
+export * from "./SplashCard"
+export * from "./Typography"
+export * from "./Card"
+export * from "./Navbar"
+export * from "./AppLayoutWrapper"
